build(order): move to Spring Boot 4.1.1 with serialization frozen
Parent 3.5.16 -> 4.1.1 with spring.jackson.use-jackson2-defaults enabled, so
this commit carries the version and the import moves while serialized output
stays byte-identical. The flag comes out in the ABSORB commit, which is what
lets a body diff found later be bisected to the defaults rather than the bump.

Three Boot 4 changes that fail silently or late rather than at compile time:

- FlywayAutoConfiguration moved into the modular spring-boot-flyway, so
  declaring only flyway-core left Flyway not running at all: no migration and
  no error. Surfaced solely because ddl-auto: validate is armed.
- The MockMvc test slice's security auto-configuration moved out of
  spring-boot-test-autoconfigure and into spring-boot-security-test, which
  contributes its own AutoConfigureMockMvc.imports. spring-security-test is
  therefore swapped for spring-boot-starter-security-test rather than dropped:
  a slice's security posture arrives through .imports, not through the test
  API, so counting @WithMockUser/springSecurity() usages cannot see it.
- Jackson 3 throws unchecked JacksonException where Jackson 2 threw IOException.
  In ProductReservationClient that is a compile error, which announces itself;
  in PaymentEventConsumer the clause simply stops matching, which does not, and
  would have sent a malformed payment event back to the queue instead of the DLQ.

The pgjdbc override stays at 42.7.12 deliberately: the driver is the mechanism
behind pod-loss readiness, so it must not move inside a version bump.

141 tests, 0 failures / 0 errors / 0 skips; per-class results identical to the
3.5.16 baseline, no class missing and no per-class count changed.

Also deletes the 3.5.16-only <netty.version> companion pin, which is part of
this bump rather than a follow-up: from 4.1.x the BOM manages netty 4.2.17,
so keeping 4.1.136 would DOWNGRADE across a minor line. The whole block goes,
comment included -- the comment quotes its own verification command, so
deleting only the property leaves `git grep -c netty.version` at 1. The
amqp-client 5.33.1 pin STAYS: 4.1.x still manages the vulnerable 5.30.0.

+++ services/order/src/test/java/com/ecommerce/order/PaymentEventListenerContainerIntegrationTest.java
@@ -194,7 +194,7 @@ class PaymentEventListenerContainerIntegrationTest extends AbstractIntegrationTe
         paymentId, orderId, USER_ID);
   }
 
-  private int readyMessages(String queue) {
+  private long readyMessages(String queue) {
     return rabbitAdmin.getQueueInfo(queue).getMessageCount();
   }
